ml_service/requirements/production.txt
Flask==1.0.2
Flask-Script==2.0.6
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
SQLAlchemy-Utils==0.33.11
PyMySQL==0.9.3
Cryptography==2.6.1
Celery==4.3.0
Redis==3.2.1